requirements.txt
# FP&A Automation Assistant - Core Dependencies
# Python 3.11+

# Core Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1

# Google Workspace Integration
gspread==6.1.2
google-auth==2.35.0
google-api-python-client==2.150.0

# Presentation & Reporting
python-pptx==1.0.2
matplotlib==3.9.2
plotly==5.24.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
typing-extensions==4.9.0
